# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: N8N AI Agent Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/workflows:
    get:
      operationId: getWorkflows
      tags: [workflows]
      summary: Get all workflows
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, all]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Workflows list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowsResponse"
  /v1/workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetailResponse"

    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/workflows/{id}/activate:
    post:
      operationId: activateWorkflow
      tags: [workflows]
      summary: Activate workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/workflows/{id}/deactivate:
    post:
      operationId: deactivateWorkflow
      tags: [workflows]
      summary: Deactivate workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/workflows/{id}/executions:
    get:
      operationId: getWorkflowExecutions
      tags: [workflows]
      summary: Get workflow executions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Executions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionsResponse"
  /v1/workflows/{id}/versions:
    get:
      operationId: getWorkflowVersions
      tags: [workflows]
      summary: Get workflow versions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Versions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionsResponse"
  /v1/workflows/bulk-action:
    post:
      operationId: bulkActionWorkflows
      tags: [workflows]
      summary: Bulk action on workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkActionRequest"
      responses:
        "200":
          description: Bulk action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Chat
...
  /v1/templates/{id}/use:
    post:
      operationId: useTemplate
      tags: [templates]
      summary: Use a template (opens chat with it)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New conversation created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"

  # Users
  /v1/users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /v1/users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/users/{id}/status:
    put:
      operationId: updateUserStatus
      tags: [users]
      summary: Toggle user active status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/users/{id}/permissions:
    put:
      operationId: updateUserPermissions
      tags: [users]
      summary: Update user permissions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionsRequest"
      responses:
        "200":
          description: Permissions updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResponse"

  # Settings
  /v1/settings/n8n/test:
    post:
      operationId: testN8nConnection
      tags: [settings]
      summary: Test n8n connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestN8nRequest"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/N8nTestResponse"
...
  /v1/settings/openai/test:
    post:
      operationId: testOpenAI
      tags: [settings]
      summary: Test OpenAI API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestApiKeyRequest"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyTestResponse"
...
  /v1/settings/gemini/test:
    post:
      operationId: testGemini
      tags: [settings]
      summary: Test Gemini API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestApiKeyRequest"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyTestResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        forcePasswordChange:
          type: boolean
        lastLogin:
          type: string
        createdAt:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
      required:
        - id
        - username
        - role
        - isActive
        - forcePasswordChange
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/User"
      required:
        - success
        - data
    UsersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: "#/components/schemas/User"
            total:
              type: integer
          required:
            - users
            - total
      required:
        - success
        - data
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
        permissions:
          type: array
          items:
            type: string
      required:
        - username
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
    UpdateUserStatusRequest:
      type: object
      properties:
        isActive:
          type: boolean
      required:
        - isActive
    UpdatePermissionsRequest:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              enabled:
                type: boolean
            required:
              - key
              - enabled
      required:
        - permissions
    ResetPasswordResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            newPassword:
              type: string
          required:
            - newPassword
      required:
        - success
        - data
    Permission:
      type: object
      properties:
        key:
          type: string
        isEnabled:
          type: boolean
      required:
        - key
        - isEnabled
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        active:
          type: boolean
        nodes:
          type: array
          items:
            type: object
            additionalProperties: true
        connections:
          type: object
          additionalProperties: true
        settings:
          type: object
          additionalProperties: true
        createdAt:
          type: string
        updatedAt:
          type: string
        successRate:
          type: number
        lastExecution:
          type: string
        executionCount:
          type: integer
        isRunning:
          type: boolean
      required:
        - id
        - name
        - active
    WorkflowsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            workflows:
              type: array
              items:
                $ref: "#/components/schemas/Workflow"
            total:
              type: integer
          required:
            - workflows
            - total
      required:
        - success
        - data
    WorkflowDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Workflow"
      required:
        - success
        - data
    Execution:
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        workflowName:
          type: string
        status:
          type: string
          enum: [success, error, running, waiting]
        startedAt:
          type: string
        stoppedAt:
          type: string
        duration:
          type: number
        errorMessage:
          type: string
      required:
        - id
        - status
        - startedAt
    ExecutionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            executions:
              type: array
              items:
                $ref: "#/components/schemas/Execution"
            total:
              type: integer
          required:
            - executions
            - total
      required:
        - success
        - data
    WorkflowVersion:
      type: object
      properties:
        id:
          type: integer
        versionNumber:
          type: integer
        changeDescription:
          type: string
        createdAt:
          type: string
        workflowJson:
          type: object
          additionalProperties: true
      required:
        - id
        - versionNumber
        - createdAt
    VersionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            versions:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowVersion"
          required:
            - versions
      required:
        - success
        - data
    BulkActionRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        action:
          type: string
          enum: [activate, deactivate, delete]
      required:
        - ids
        - action
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [create, edit, diagnose, query]
        status:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        userId:
          type: integer
        username:
          type: string
      required:
        - id
        - title
        - type
        - createdAt
    ConversationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Conversation"
      required:
        - success
        - data
...
    TestN8nRequest:
      type: object
      properties:
        url:
          type: string
        apiKey:
          type: string
      required:
        - url
        - apiKey
...
    N8nTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connected:
              type: boolean
            version:
              type: string
            workflowsCount:
              type: integer
            message:
              type: string
          required:
            - connected
      required:
        - success
        - data
...
    TestApiKeyRequest:
      type: object
      properties:
        apiKey:
          type: string
      required:
        - apiKey
...
    ApiKeyTestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            valid:
              type: boolean
            modelAvailable:
              type: boolean
            balanceInfo:
              type: string
            dailyQuota:
              type: string
            message:
              type: string
          required:
            - valid
      required:
        - success
        - data
...